move io to separate module, move data.pymatgen to io
